Refactor using models for easy testability
