+++ functions/backfill/post/insertion_test.go
@@ -4,14 +4,14 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/KamWithK/exSTATic-backend/utils"
+	"github.com/KamWithK/exSTATic-backend/models"
 	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
 	"github.com/jaswdr/faker"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestNull(t *testing.T) {
-	result, err := HandleRequest(nil, utils.BackfillArgs{})
+	result, err := HandleRequest(nil, models.BackfillArgs{})
 
 	assert.Nil(t, result, "No input => no writes")
 	assert.Error(t, err)
@@ -20,9 +20,9 @@ func TestNull(t *testing.T) {
 func TestNoUsername(t *testing.T) {
 	fake := faker.New()
 
-	result, err := HandleRequest(nil, utils.BackfillArgs{
+	result, err := HandleRequest(nil, models.BackfillArgs{
 		Username:     "",
-		MediaEntries: utils.RandomMediaEntries(fake, "", 3),
+		MediaEntries: models.RandomMediaEntries(fake, "", 3),
 	})
 
 	assert.Nil(t, result, "Writes can't be performed when no username is entered")
@@ -36,9 +36,9 @@ func TestMultipleUsernames(t *testing.T) {
 
 	invalidEntries, validEntries := 5, 3
 
-	results, err := HandleRequest(nil, utils.BackfillArgs{
+	results, err := HandleRequest(nil, models.BackfillArgs{
 		Username:     user1,
-		MediaEntries: append(utils.RandomMediaEntries(fake, user2, invalidEntries), utils.RandomMediaEntries(fake, user1, validEntries)...),
+		MediaEntries: append(models.RandomMediaEntries(fake, user2, invalidEntries), models.RandomMediaEntries(fake, user1, validEntries)...),
 	})
 
 	assert.Len(t, results.WriteRequests, validEntries, "Entries with different usernames aren't valid")
@@ -49,9 +49,9 @@ func TestWriteMediaEntries(t *testing.T) {
 	fake := faker.New()
 	user := fake.Person().Name()
 
-	inputMediaEntries, producedMediaEntries := utils.RandomMediaEntries(fake, user, 100), []utils.UserMediaEntry{}
+	inputMediaEntries, producedMediaEntries := models.RandomMediaEntries(fake, user, 100), []models.UserMediaEntry{}
 
-	results, err := HandleRequest(nil, utils.BackfillArgs{
+	results, err := HandleRequest(nil, models.BackfillArgs{
 		Username:     user,
 		MediaEntries: inputMediaEntries,
 	})
@@ -60,14 +60,14 @@ func TestWriteMediaEntries(t *testing.T) {
 	assert.NotNil(t, results)
 
 	for _, writeRequest := range results.WriteRequests {
-		intermediateItem := utils.IntermediateEntryItem{}
+		intermediateItem := models.IntermediateEntryItem{}
 
 		unmarshalErr := dynamodbattribute.UnmarshalMap(writeRequest.PutRequest.Item, &intermediateItem)
 		assert.NoError(t, unmarshalErr)
 
 		splitPK := strings.Split(intermediateItem.PK, "#")
 		assert.Len(t, splitPK, 2, "PK should precisely be composed of the media type and username")
-		intermediateItem.Key = utils.UserMediaKey{
+		intermediateItem.Key = models.UserMediaKey{
 			Username:        splitPK[1],
 			MediaType:       splitPK[0],
 			MediaIdentifier: intermediateItem.SK,
@@ -83,14 +83,14 @@ func TestWriteMediaStats(t *testing.T) {
 	fake := faker.New()
 	user := fake.Person().Name()
 
-	mediaEntries := utils.RandomMediaEntries(fake, user, 100)
-	inputMediaStats, producedMediaStats := []utils.UserMediaStat{}, []utils.UserMediaStat{}
+	mediaEntries := models.RandomMediaEntries(fake, user, 100)
+	inputMediaStats, producedMediaStats := []models.UserMediaStat{}, []models.UserMediaStat{}
 
 	for _, mediaEntry := range mediaEntries {
-		inputMediaStats = append(inputMediaStats, utils.RandomMediaStats(fake, mediaEntry.Key, 30, 0.8)...)
+		inputMediaStats = append(inputMediaStats, models.RandomMediaStats(fake, mediaEntry.Key, 30, 0.8)...)
 	}
 
-	results, err := HandleRequest(nil, utils.BackfillArgs{
+	results, err := HandleRequest(nil, models.BackfillArgs{
 		Username:   user,
 		MediaStats: inputMediaStats,
 	})
@@ -99,7 +99,7 @@ func TestWriteMediaStats(t *testing.T) {
 	assert.NotNil(t, results)
 
 	for _, writeRequest := range results.WriteRequests {
-		intermediateItem := utils.IntermediateStatItem{}
+		intermediateItem := models.IntermediateStatItem{}
 
 		unmarshalErr := dynamodbattribute.UnmarshalMap(writeRequest.PutRequest.Item, &intermediateItem)
 		assert.NoError(t, unmarshalErr)
@@ -108,7 +108,7 @@ func TestWriteMediaStats(t *testing.T) {
 		splitSK := strings.Split(intermediateItem.SK, "#")
 		assert.Len(t, splitPK, 2, "PK should precisely be composed of the media type and username")
 		assert.Len(t, splitSK, 2, "SK should precisely be composed of the zero padded unix epoch date and media identifier")
-		intermediateItem.Key = utils.UserMediaKey{
+		intermediateItem.Key = models.UserMediaKey{
 			Username:        splitPK[1],
 			MediaType:       splitPK[0],
 			MediaIdentifier: splitSK[1],
