+++ functions/models/random_data_model.go
@@ -1,4 +1,4 @@
-package utils
+package models
 
 import (
 	"math/rand"
