+++ functions/settings/get/main.go
@@ -2,16 +2,12 @@ package main
 
 import (
 	"context"
-	"errors"
 
-	"github.com/KamWithK/exSTATic-backend/utils"
-	"github.com/rs/zerolog/log"
+	"github.com/KamWithK/exSTATic-backend/models"
 
 	"github.com/aws/aws-lambda-go/lambda"
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
-	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
 )
 
 var sess *session.Session
@@ -24,35 +20,8 @@ func init() {
 	svc = dynamodb.New(sess)
 }
 
-func HandleRequest(ctx context.Context, key utils.UserSettingsKey) (*utils.UserSettings, error) {
-	tableKey, keyErr := dynamodbattribute.MarshalMap(key)
-	if keyErr != nil {
-		log.Error().Err(keyErr).Str("table", "settings").Interface("key", key).Msg("Could not unmarshal dynamodb key")
-		return nil, keyErr
-	}
-
-	result, getErr := svc.GetItem(&dynamodb.GetItemInput{
-		TableName: aws.String("settings"),
-		Key:       tableKey,
-	})
-	if getErr != nil {
-		log.Error().Str("table", "settings").Interface("key", key).Msg("Dynamodb failed to get item")
-		return nil, getErr
-	}
-
-	if result.Item == nil || len(result.Item) == 0 {
-		log.Info().Str("table", "settings").Interface("key", key).Msg("Item not in table")
-		return nil, errors.New("Item not found in table")
-	}
-
-	optionArgs := utils.UserSettings{}
-	if unmarshalErr := dynamodbattribute.UnmarshalMap(result.Item, &optionArgs); unmarshalErr != nil {
-		log.Error().Err(unmarshalErr).Str("table", "settings").Interface("key", key).Interface("item", result.Item).Msg("Could not unmarshal dynamodb item")
-		return nil, unmarshalErr
-	}
-	optionArgs.Key = key
-
-	return &optionArgs, nil
+func HandleRequest(ctx context.Context, key models.UserSettingsKey) (*models.UserSettings, error) {
+	return models.GetUserSettings(svc, key)
 }
 
 func main() {
